Refactor migration scripts and add average bill by last month usecase

+++ src/main/java/net/example/coffeeshop/entrypoints/http/controllers/response/AverageBillByLastMonthResponse.java
@@ -3,6 +3,8 @@ package net.example.coffeeshop.entrypoints.http.controllers.response;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Builder
 public class AverageBillByLastMonthResponse {
 
@@ -11,5 +13,5 @@ public class AverageBillByLastMonthResponse {
     private Long shopId;
 
     @Getter
-    private Long avgBill;
+    private BigDecimal avgBill;
 }
